tambahkan CRUD

+++ src/test/java/ui/ft/ccit/faculty/transaksi/barang/BarangRepositoryTest.java
@@ -26,7 +26,7 @@ class BarangRepositoryTest {
                                 5000,
                                 20.0,
                                 0.0,
-                                103,
+                                (byte) 103,
                                 "S003");
 
                 barangRepository.save(barang);
@@ -39,8 +39,8 @@ class BarangRepositoryTest {
 
         @Test
         void findByNamaContainingIgnoreCase_shouldReturnMatchingRows() {
-                barangRepository.save(new Barang("T001", "Sabun Mandi", 10, 5000, 20.0, 0.0, 103, "S003"));
-                barangRepository.save(new Barang("T002", "Shampoo Wangi", 5, 15000, 25.0, 0.0, 103, "S003"));
+                barangRepository.save(new Barang("T001", "Sabun Mandi", 10, 5000, 20.0, 0.0, (byte) 103, "S003"));
+                barangRepository.save(new Barang("T002", "Shampoo Wangi", 5, 15000, 25.0, 0.0, (byte) 103, "S003"));
 
                 List<Barang> hasil = barangRepository.findByNamaContainingIgnoreCase("sham");
